refactor: isolate shared emulator services
