+++ crates/dingooemu-core/src/a320/mod.rs
@@ -1,3 +1,4 @@
+mod cheats;
 pub mod cpu;
 mod diagnostics;
 #[cfg(feature = "jit")]
