+++ crates/dingooemu-core/src/a330/mod.rs
@@ -1,3 +1,4 @@
+mod cheats;
 pub mod cpu;
 mod firmware_archive;
 pub mod memory;
