+++ crates/dingooemu-core/src/common/mod.rs
@@ -1,2 +1,7 @@
+pub mod audio;
+pub mod cheats;
 pub(crate) mod execution;
 pub(crate) mod hle;
+pub mod input;
+pub(crate) mod save_state;
+pub mod video;
