+++ crates/dingooemu-core/src/lib.rs
@@ -6,21 +6,16 @@
 
 pub mod a320;
 pub mod a330;
-pub mod audio;
-pub mod cheats;
-mod common;
+pub mod common;
 pub mod content;
+mod emulator;
 pub mod error;
-pub mod input;
 pub mod package;
-mod runtime;
-mod save_state;
-pub mod video;
 
 // Re-export main types for convenience
 pub use a320::JitDiagnostics;
 pub use common::execution::UnknownInstructionPolicy;
 pub use common::hle::{UnknownHleCall, UnknownHlePolicy};
 pub use content::{ArmProfile, ContentFormat, GuestArchitecture, TargetDevice};
+pub use emulator::Emulator;
 pub use error::{Result, SimulatorError};
-pub use runtime::Emulator;
