+++ crates/dingooemu-libretro/src/api.rs
@@ -4,12 +4,12 @@ use std::ptr;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Instant;
 
-use dingooemu_core::audio::OUTPUT_SAMPLE_RATE;
-use dingooemu_core::input::{
+use dingooemu_core::common::audio::OUTPUT_SAMPLE_RATE;
+use dingooemu_core::common::input::{
     BUTTON_A, BUTTON_B, BUTTON_DOWN, BUTTON_L, BUTTON_LEFT, BUTTON_R, BUTTON_RIGHT, BUTTON_SELECT,
     BUTTON_START, BUTTON_UP, BUTTON_X, BUTTON_Y,
 };
-use dingooemu_core::video::{SCREEN_HEIGHT, SCREEN_WIDTH};
+use dingooemu_core::common::video::{SCREEN_HEIGHT, SCREEN_WIDTH};
 use dingooemu_core::Emulator;
 use dingooemu_core::UnknownInstructionPolicy;
 
@@ -1097,7 +1097,7 @@ mod tests {
         assert_eq!(SYSTEM_RAM_MAP_START.load(Ordering::SeqCst), 0);
         assert_eq!(
             VIDEO_RAM_MAP_START.load(Ordering::SeqCst),
-            dingooemu_core::video::VM_LCD_FB_ADDRESS as usize
+            dingooemu_core::a320::memory::LCD_FRAMEBUFFER_BASE as usize
         );
         assert_eq!(
             retro_get_memory_size(RETRO_MEMORY_SYSTEM_RAM),
