+++ crates/dingooemu/src/gamepad_overlay.rs
@@ -1,4 +1,4 @@
-use dingooemu_core::input::{
+use dingooemu_core::common::input::{
     BUTTON_A, BUTTON_B, BUTTON_DOWN, BUTTON_L, BUTTON_LEFT, BUTTON_R, BUTTON_RIGHT, BUTTON_SELECT,
     BUTTON_START, BUTTON_UP, BUTTON_X, BUTTON_Y,
 };
